contrib/io_mapping/requirements.txt
argcomplete==1.11.1
asn1crypto==1.3.0
autoclick==0.8.0
beautifulsoup4==4.4.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
coloredlogs==10.0
cryptography==2.2.2
docker==4.2.0
docparse==0.2.0
dxpy==0.290.1
entrypoints==0.3
flake8==3.7.9
humanfriendly==6.1
idna==2.8
importlib-metadata==1.5.0
lark-parser==0.7.8
mccabe==0.6.1
miniwdl==0.6.4
ngs==1.0
psutil==5.6.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pygtail==0.11.1
python-dateutil==2.8.1
python-magic==0.4.6
regex==2020.1.8
requests==2.22.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.53.0
zipp==2.2.0
